pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyargdown"
dynamic = ["version"]
description = 'Parsing and processing argdown snippets in python'
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0"
keywords = []
authors = [
  { name = "Gregor Betz", email = "3662782+ggbetz@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "lark>=1.2.2",
    "networkx>=3.4.2,<3.6",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "datasets",
]

[project.urls]
Documentation = "https://github.com/debatelab/pyargdown#readme"
Issues = "https://github.com/debatelab/pyargdown/issues"
Source = "https://github.com/debatelab/pyargdown"

[tool.hatch.version]
path = "src/pyargdown/__about__.py"

[tool.hatch.envs.test]
python = "3.10"
dependencies = [
  "pytest>=8.3.4",
  "datasets",
]
[tool.hatch.envs.test.scripts]
run = "pytest {args:tests/}"

